List and Alert
List and Alert

+++ app/src/main/java/com/example/listbuyapp/MainActivity.kt
@@ -1,19 +1,22 @@
 package com.example.listbuyapp
 
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import android.view.View
 import android.widget.TextView
 import androidx.appcompat.app.ActionBarDrawerToggle
+import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import androidx.core.content.ContextCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
 import com.example.listbuyapp.databinding.ActivityMainBinding
 import com.google.firebase.auth.FirebaseAuth
+import com.squareup.picasso.Picasso
+import de.hdodenhof.circleimageview.CircleImageView
 
 enum class ProviderType {
     BASIC,
@@ -65,14 +68,36 @@ class MainActivity : AppCompatActivity() {
                 R.id.history -> replaceFragment(historyFragment(), item.title.toString())
                 R.id.categories -> replaceFragment(categoriesFragment(), item.title.toString())
                 R.id.logOut -> {
-                    val prefs = getSharedPreferences(
-                        getString(R.string.prefs_file),
-                        Context.MODE_PRIVATE
-                    ).edit()
-                    prefs.clear()
-                    prefs.apply()
-                    FirebaseAuth.getInstance().signOut()
-                    onBackPressed()
+
+
+                    val alertDialogBuilder = AlertDialog.Builder(this)
+                    alertDialogBuilder.setTitle("Desea cerrar sesión?")
+                    alertDialogBuilder.setMessage("Cerrar")
+                    alertDialogBuilder.setPositiveButton("Si") { dialog, _ ->
+                        dialog.dismiss()
+                        val prefs = getSharedPreferences(
+                            getString(R.string.prefs_file),
+                            Context.MODE_PRIVATE
+                        ).edit()
+                        prefs.clear()
+                        prefs.apply()
+                        FirebaseAuth.getInstance().signOut()
+                        startActivity(Intent(this, Login::class.java))
+                        finish()
+                    }
+                    alertDialogBuilder.setNegativeButton("No") { dialog, _ ->
+                        dialog.dismiss()
+                    }
+
+                    val alertDialog = alertDialogBuilder.create()
+                    alertDialog.show()
+
+                    val positiveButton = alertDialog.getButton(androidx.appcompat.app.AlertDialog.BUTTON_POSITIVE)
+                    positiveButton.setTextColor(ContextCompat.getColor(this, R.color.botonbase))
+                    val negativeButton = alertDialog.getButton(androidx.appcompat.app.AlertDialog.BUTTON_NEGATIVE)
+                    negativeButton.setTextColor(ContextCompat.getColor(this, R.color.botonbase))
+
+                    alertDialog.window?.setBackgroundDrawableResource(R.drawable.background_dialog)
                 }
             }
             true
@@ -82,10 +107,16 @@ class MainActivity : AppCompatActivity() {
         val emailc: TextView = headerView.findViewById(R.id.emailTextView)
         val namec: TextView = headerView.findViewById(R.id.nameTextView)
 
-        val name = firebaseUser?.displayName.takeIf { it?.isNotEmpty() == true } ?: "User"
+        if (firebaseUser != null && firebaseUser.photoUrl != null) {
+            val photoUrl: String = firebaseUser.photoUrl.toString()
+            val imageView: CircleImageView = headerView.findViewById(R.id.userImageView)
+
+            Picasso.get().load(photoUrl).into(imageView)
+        }
+
+        val name = firebaseUser?.displayName.takeIf { it?.isNotEmpty() == true } ?: "!Hola¡ Usuario"
         namec.text = name
         emailc.text = email
-
         replaceFragment(homeFragment(), "Home")
     }
